move surface::text over to text.cpp
