+++ 32blit/graphics/surface.cpp
@@ -446,78 +446,6 @@ namespace blit {
   /**
    * TODO: Document this function
    *
-   * \param message
-   * \param font
-   * \param p
-   * \param variable
-   */
-  void surface::text(std::string message, const uint8_t *font, const point &p, bool variable) {
-    text(message, font, rect(p.x, p.y, 10000, 10000), variable);
-  }
-
-  /**
-   * TODO: Document this function
-   *
-   * \param message
-   * \param font
-   * \param r
-   * \param variable
-   */
-  void surface::text(std::string message, const uint8_t *font, const rect &r, bool variable) {
-    point c(r.x, r.y); // caret position
-
-    for (char &chr : message) {
-      // draw character
-      
-      uint8_t chr_idx = chr & 0x7F;
-      chr_idx = chr_idx < ' ' ? 0 : chr_idx - ' ';
-
-      uint8_t char_width = 0;
-      
-      const uint8_t* font_chr = &font[chr_idx * 6];
-      
-      for (uint8_t y = 0; y < 8; y++) {          
-        uint32_t po = offset(point(c.x, c.y + y));
-
-        for (uint8_t x = 0; x < 6; x++) {
-            
-          //if(clip.contains(p)) {
-            if (font_chr[x] & (1 << y)) {
-              bf((uint8_t *)&_pen, this, po, 1);
-              char_width = char_width < x ? x : char_width;
-            }
-          //}
-
-          //p.x++;
-          po++;
-        }
-
-        //p.y++;
-      }
-
-      if (!variable)
-        char_width = 4;
-
-      char_width += 2;
-
-      if (chr == 32) {
-        char_width = 3;
-      }
-
-      // increment the cursor
-      c.x += char_width;
-      if ((c.x > r.x + r.w) | (chr == 10)) {
-        c.x = r.x;
-        c.y += 9;
-      }
-    }
-
-    //return c.y + 8;
-  }
-
-  /**
-   * TODO: Document this function
-   *
    * \param image
    */
   void surface::load_from_packed(const packed_image *image) {
